requirements.txt
# Dependencias principales
opencv-python==4.8.0.76
numpy

# MTCNN para detección de rostros (no requiere compilación)
mtcnn==0.1.1

# TensorFlow (versión compatible con Python 3.11)
tensorflow==2.15.0

# Keras y modelos preentrenados
keras==2.15.0

# Face recognition SIN dlib (alternativa más fácil)
deepface==0.0.79

# Pillow para procesamiento de imágenes
Pillow==10.0.0

# Asistente IA con Groq
groq==0.34.1
python-dotenv==1.0.0

# Interfaz web
streamlit==1.28.0
gradio==4.0.0

# Utilidades
pandas==2.0.3
matplotlib==3.7.2
scikit-learn==1.3.0